eth_call tests
