+++ packages/eth-rpc-adapter/src/__tests__/endpoint-tests/endpoint.ts
@@ -610,90 +610,6 @@ describe('endpoint', () => {
   });
 
 
-  describe('eth_call', () => {
-    const callRequest = (abi: any) => async (address: string, method: string, params?: any[], blockTag?: any) => {
-      const iface = new Interface(abi);
-
-      const data = iface.encodeFunctionData(method, params);
-      const block = blockTag || (await eth_blockNumber()).data.result;
-      const rawRes = (await eth_call([{ to: address, data }, block])).data.result;
-
-      return iface.decodeFunctionResult(method, rawRes);
-    };
-
-    const callToken = callRequest(TokenABI.abi);
-    const callDex = callRequest(DEXABI.abi);
-
-    it('get correct procompile token info', async () => {
-      // https://github.com/AcalaNetwork/Acala/blob/a5d9e61c74/node/service/src/chain_spec/mandala.rs#L628-L636
-      // Native tokens need to registry in asset_registry module.
-      const tokenMetaData = [
-        {
-          address: '0x0000000000000000000100000000000000000000',
-          name: 'Acala',
-          symbol: 'ACA',
-          decimals: 12,
-        },
-        {
-          address: '0x0000000000000000000100000000000000000001',
-          name: 'Acala Dollar',
-          symbol: 'AUSD',
-          decimals: 12,
-        },
-        {
-          address: '0x0000000000000000000100000000000000000002',
-          name: 'Polkadot',
-          symbol: 'DOT',
-          decimals: 10,
-        },
-        {
-          address: '0x0000000000000000000100000000000000000003',
-          name: 'Liquid DOT',
-          symbol: 'LDOT',
-          decimals: 10,
-        },
-      ];
-
-      const tests = tokenMetaData.map(async ({ address, name, symbol, decimals }) => {
-        const [_name] = await callToken(address, 'name');
-        const [_symbol] = await callToken(address, 'symbol');
-        const [_decimals] = await callToken(address, 'decimals');
-
-        expect(_name).to.equal(name);
-        expect(_symbol).to.equal(symbol);
-        expect(_decimals).to.equal(decimals);
-      });
-
-      await Promise.all(tests);
-    });
-
-    it('supports calling historical blocks', async () => {
-      const dexAddr = '0x0230135fded668a3f7894966b14f42e65da322e4'; // created at block 5
-      const before = await callDex(dexAddr, 'getLiquidityPool', [ADDRESS.ACA, ADDRESS.AUSD], { blockNumber: '0x5' });
-      // swap happens at block 6
-      const block7Hash = (await eth_getBlockByNumber([7, false])).data.result.hash;
-      const after = await callDex(dexAddr, 'getLiquidityPool', [ADDRESS.ACA, ADDRESS.AUSD], { blockHash: block7Hash });
-
-      expect(before.map(BigInt)).to.deep.equal([1000000000000000000n, 2000000000000000000n]);
-      expect(after.map(BigInt)).to.deep.equal([1000002000000000000n, 1999996004007985992n]);
-    });
-
-    it('throws correct error for invalid tag', async () => {
-      const dexAddr = '0x0230135fded668a3f7894966b14f42e65da322e4';
-      const data = '0x123123123';
-
-      expect((await eth_call([{ to: dexAddr, data }, { hahaha: 13542 }])).data.error).to.deep.equal({
-        code: -32602,
-        message: 'invalid argument 1: invalid eip-1898 blocktag, expected to contain blockNumber or blockHash',
-      });
-
-      expect((await eth_call([{ to: dexAddr, data }, { blockHash: 123 }])).data.error).to.deep.equal({
-        code: -32602,
-        message: 'invalid argument 1: invalid block hash, expected type String',
-      });
-    });
-  });
-
   describe('eth_getCode', () => {
     const preCompileAddresses = [
       '0x0000000000000000000100000000000000000001', // AUSD
